+++ src/main/java/com/nps/sdk/java/response/ClientListResult.java
@@ -52,9 +52,9 @@ public class ClientListResult implements BaseResponse {
 		@Data
 		public static class Flow {
 
-			private int ExportFlow;
+			private long ExportFlow;
 			private long InletFlow;
-			private int FlowLimit;
+			private long FlowLimit;
 		}
 
 		@Data
